item transfer
Got the scene to deload and the player and such became a child of the new scene, but i couldnt interact with anything. Decided to change to having all the scenes in 1 and teleporting between them. Tested teleporting with items in hand and pockets and they come too so less work for me!